projects/speaker-identification/requirements.txt
torch>=2.0.0
torchaudio>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
scipy>=1.10.0
sounddevice>=0.4.6
librosa>=0.10.0
matplotlib>=3.7.0
scikit-learn>=1.3.0
flask>=2.3.0
flask-cors>=4.0.0
noisereduce>=2.0.0
pillow>=10.0.0
